test: data 폴더 -> __mock__ 폴더 로 이동
