+++ server/apps/api/test/user.e2e-spec.ts
@@ -9,9 +9,9 @@ import { initTestUser1, initTestUser2, user1Modify } from '@mock/user.mock';
 import { importConfigModule } from '@api/modules/Config.module';
 import { importWinstonModule } from '@api/modules/Winstone.module';
 import { followingURL, getMyInfoURL, signupURL, signinURL } from '@api/test/urls/urls';
-import { userData } from '@api/test/data/userData';
 import { ApiInterceptor } from '@custom/interceptor/api.interceptor';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
+import { userData } from '@mock/auth.mock';
 
 describe('User E2E Test', () => {
   let app, server, userModel, mongod, user1;
